Update Diguardi_BCI.Rmd
I did some of the reading I had stacked up and tried to get it to correlate to my subject matter. I found that there would be more information front he books that I added to the bibtext. 
I contacted Neruable, they are focusing on the head gear for public use not the VR controller. I think they were the closest to the idea I had to continue after school.
Overall I was focusing on other schoolwork and the borderlands reading over this paper. I feel behind.